components/db/pyproject.toml
[project]
name = "pinta-db"
version = "0.1.0"
description = "Database component of the elevation production system"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "geoalchemy2[shapely]>=0.18.4",
  "sqlmodel>=0.0.37",
]

[tool.uv]
default-groups = ["dev", "lint"]

[dependency-groups]
dev = [
  "alembic>=1.18.4",
  "dotenv>=0.9.9",
  "psycopg[binary]>=3.3.3",
  "pytest-dotenv>=0.5.2",
  "pytest-xdist>=3.8.0",
]
lint = [
  "prek>=0.3.5",
]

[tool.uv.sources]
pinta-test-utils = { workspace = true }

[build-system]
requires = ["setuptools>=82", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pinta_db = ["py.typed"]

[tool.ruff]
# Extend the root `pyproject.toml`
extend = "../../pyproject.toml"

[tool.ruff.lint.extend-per-file-ignores]
"*/versions/*" = [
  "D", # pydocstyle

]

[tool.pytest.ini_options]
addopts = "--import-mode=importlib -n auto"
